msys2-runtime: upgrade to v3.4.6
This corresponds to
https://github.com/git-for-windows/msys2-runtime/pull/50.

The `update-patches.sh` script had to be adjusted to reflect that Cygwin
changed its tag naming scheme. Before, it was `cygwin-<version>-release`
where the version had underscores instead of dots. Now, it is
`cygwin-<version>` with dots, not underscores. Also, the source URL
changed to `git+https` but we obviously cannot clone from such a URL.

Finally, the build no longer produces an `msys0.dll`, but a
`new-msys-2.0.dll` and the Git for Windows-only code-signing step had to
be adjusted accordingly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>