msys2-runtime: fix code-signing
The problem with code-signing is that `strip.exe` does something
apparently awful to the code-signature.

Of course, stripping a binary _has_ to invalidate the signature,
otherwise what would be the point of code-signing? But `strip.exe` does
something much worse, something that lets `osslsigncode` error out with:

	Corrupt PE file - current signature not at end of file

And it does not matter whether the binary is _actually_ signed. Even
when osslsigncode is asked to sign an unsigned & stripped `.dll`, it
errors out with that message.

This was a problem earlier, too, which is why I had to be very careful
_not_ to enable the `strip` option.

But when I upgraded to MSYS2 runtime v3.4.6, the binary grew from 3.4MB
to 23MB, so I did enable it. And broke code-signing.

The reason for this hike in size is that Cygwin no longer splits out the
debug information into a separate `.dbg` file, and hence nowadays it is
all still in the `.dll`.

Let's just not produce such big binaries in the first place, and avoid
stripping the `.dll` files again. And thereby unbreak code-signing.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>