msys2-runtime: adapt `update-patches.sh` to the new world order
Since upgrading to v3.4.6, the layout of the section that applies the
patches is different.

As I had forgotten to update that, and as I had missed that it had not
updated that section at all, msys2-runtime-3.4.6-2 did not actually
apply the patch that was added since msys2-runtime-3.4.6-1. D'oh!

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>