msys2-runtime: _actually_ fix Git's t3701-add-interactive
In https://github.com/git-for-windows/MSYS2-packages/pull/104, I tried
to fix a hang by reverting a problematic commit. This seemed to work, as
the patch was included, but I failed to notice that the
`update-patches.sh` script had to be adjusted so that the patch _would
also be applied_.

After the previous commit fixed the `update-patches.sh` script, this is
now the fix-up commit to actually apply said patch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>