openssh/curl: bump pkgrel to rebuild after OpenSSL upgrade _again_
OpenSSL was rebuilt, but without a valid `libssl.dll.a` to use,
therefore the i686 variant of the `curl` package could not be built (and
the `x86_64` variant could only be built in a bogus version that
unintentionally links _statically_ to OpenSSL).

In preparation for a fixed OpenSSL, we bump the pkgrel of the dependees
that will need to be built again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>