serf: sync with MSYS2
The Git for Windows project does not usually maintain this package but
simply consumes it from the MSYS2 project. However, the Git for Windows
project needs to support i686, still, which MSYS2 has dropped.
Therefore, we will have to rebuild this package because one of its
dependencies, OpenSSL, was upgraded in a way that changes the respective
DLLs' names.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>