msys2-runtime: handle "untracked" file correctly
The `getprocaddr.c` file is added by an add-on patch. Therefore, when
rebuilding the package the patch does not apply because that file
already is there.

This is very similar to the situation with the `msys2_path_conv.*` and
`exit_process.h` files, therefore let's treat it the same.

This fixes https://github.com/git-for-windows/git/issues/4442

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>