msys2-runtime: avoid using absolute alternative object directory paths
When `makepkg` initializes a source that refers to a remote Git
repository, it first makes a bare clone, and then initializes the
worktree in `src/<name>` with that bare clone as a `--reference`.

Internally, this asks Git to use the Git objects from that bare clone by
installing an `alternates` file that contains paths. Absolute paths.

And that is a problem because on Windows, absolute paths are something
different than what the MSYS2 runtime considers absolute. Therefore,
since `makepkg` uses an _MSYS_ variant of `git.exe`, these paths are
Unix-style absolute paths, starting with a forward slash, and really
being relative to the MSYS2 root directory. When accessing the worktree
via a _MINGW_ variant of `git.exe`, which has no clue about pseudo-Unix
paths, it says that it cannot find that alternate because it expected
_Windows-style_ absolute directories, i.e. starting with a drive letter
and a colon. But using a colon in that path would confuse _MSYS_
`git.exe`, therefore we simply cannot use absolute paths.

So let's replace them by relative ones.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>