msys2-runtime: fix git-for-windows CI failures
In Git for Windows' CI runs, `tar` is used to unpack an archive that
contains a symbolic link. This had been broken in Cygwin runtime v3.4.7,
and here is a fix.

While at it, also fix the regression where the `msys2-runtime` commit
hash was no longer mentioned in the output of `uname -r`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>