msys2-runtime: restore fast path for current user primary group
A recent commit removed an optimization from the Cygwin runtime that
fetches the default group from the current user token: At times, this is
not accurate such as when groups like the builtin Administrators group
is the primary group.

However, removing this optimization causes extremely poor
performance when connected to some Active Directory
environments.

Reflecting https://github.com/git-for-windows/msys2-runtime/pull/57,
this update restores this optimization as the default behaviour. To
still allow the user to opt into the correct behavior, a `group:
db-accurate` option is provided for `nsswitch.conf`.

This fixes https://github.com/git-for-windows/git/issues/4459

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>