msys2-runtime: fix diff context
One of the recent patches applies cleanly when cherry-picked, but not
when applied as a patch, because I merged the `msys2-3.4.9` branch whose
patches' context is changed by patches that are in Git for Windows only,
and that are applied before the fixups.

I cherry-picked the last patch manually, generated the diff, and
adjusted the patch. This will have to do for now, if we run
`./update-patches.sh` again before rebasing to a newer MSYS2 runtime
version, we will have to re-apply this here commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>