libksba: synchronize with MSYS2
The libksba version 1.6.3 or newer is required by the most recent GNU
Privacy Guard release train, therefore we need to update this package in
git-sdk-32.

This synchronizes the package definition with f3e0d870 (libksba: Update
to 1.6.5, 2023-11-18), carefully leaving out the following changes:

- pretending to be Cygwin (because it does not work with the MSYS2
  runtime we have in git-sdk-32)
- closely related: dropping the msysize patch
- depending on `autotools` (because that was established only after
  MSYS2 dropped i686 support)

The msysize.patch file had to be adjusted to the new version, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>