ci: for PR's whose title start with `i686:` don't build x86_64
Git for Windows will run increasingly into the situation where packages
that are somewhat outside of Git for Windows' area of interest (such as
`libksba`) have to be built for i686, solely because they are outdated
dependencies of packages Git for Windows _wants_ to build (such as
`gnupg`).

So far, this is a pretty manual process, but since more and more
packages will fall into this bucket, let's start automating things
better.

With this commit, we skip the x86_64 CI builds from PRs whose title
starts with `i686:`, so that we can label clearly what needs to be
tested only on i686 (because it will only be deployed to git-sdk-32).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>