msys2-runtime: backport a couple of console-related fixes from v3.5.0
There have been a couple of fixes for hangs recently, sadly only for
v3.5.0, but most of them apply to v3.4.x, too. So let's backport them.

An alternative would be to switch to v3.5.0, but Git for Windows is in
its last days before v2.44.0 is scheduled to be released, so we'll hold
off from making that jump already. It's not only the closeness of a new
official Git for Windows version, though:

- The upgrade to MSYS2 runtime v3.5.0 would have huge implications:
  Windows 7/8 support will be dropped.

- I immediately hit a bug in Cygwin v3.5.0, where calling
  (non-Cygwin/MSYS) git.exe from the MSYS2 Bash would always report
  success, even if the exit code of the git.exe process was non-zero.

- Cygwin v3.5.0 has been out for just two weeks, and given the severity
  of the already-found bug, it seems prudent to give it much more time
  to show (and resolve) bugs.

- There is still this vexing MSYS2 runtime hang in our Windows/ARM64
  pipeline, it happens even with MSYS2 runtime v3.5.0. Read: those hangs
  are not even fixed by this here commit. But other hangs are fixed, so
  it's a good change nevertheless.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>