msys2-runtime: restore blocking mode of read pipe on close()
The Cygwin runtime (and for that reason, the MSYS2 runtime, too), change
pipes from blocking to non-blocking, which is a problem when the pipe is
actually opened by something like Git when it asks a credential helper
for information. Git, not expecting the pipe mode to be changed behind
its back, will then error out with "fatal: read error: Invalid
argument".

To address this, this patch was proposed in the Cygwin project as
https://inbox.sourceware.org/cygwin-patches/20240830141553.12128-1-takashi.yano@nifty.ne.jp/

Ideally, Cygwin would not even _try_ to fiddle with the pipe mode,
therefore this patch was not applied as-is. At time of writing, there is
no consensus on any replacement patch, but I have to prepare _something_
for the already-late Git for Windows v2.46.1 (which is likely the last
Git for Windows version to support Windows 7 and Windows 8, hence the
urgency). And this patch at least improves the situation.

This patch will be dropped when Git for Windows will upgrade to MSYS2
runtime v3.5, and hopefully consensus will have been reached about a
better fix by that time.

This commit corresponds to
https://github.com/git-for-windows/msys2-runtime/pull/72 which fixes
https://github.com/git-for-windows/git/issues/5115.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>