msys2-runtime(update-patches): generate the canonical stable checksums
In 111966f097 (msys2-runtime(update-patches.sh): do generate stable
checksums for the Cygwin code, 2024-07-08), I identified a problem that
needed fixing: `updpkgsums` would now use the SHA-256 of the tar archive
generated from the Cygwin tag as the checksum for the first `source`
array item. But that checksum would depend on the value of
`core.autoCRLF`!

Turns out that our CI uses the exact opposite value of the one I
hard-coded. So let's fix that now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>