ci: work around libcurl v8.11.0 problem
With libcurl v8.11.0, `git fetch` fails (curiously only with the MSYS
version, the MINGW version seems to have no problems in Git for Windows'
CI builds).

This would cause problems e.g. in the `CI-Build` step of the PR builds
in MSYS2-packages, which by itself would prevent an upgrade of libcurl
to a working version! Catch 22.

Work around that by detecting the situation and force-downgrading in
that case.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>