pcre2: update to 10.45
This version was released by a new maintainer (and hence was signed with
a new PGP key), for details, see
https://groups.google.com/g/pcre2-dev/c/eJgfSV-faVY/m/ev5VJEtyAAAJ.

It also is the first release where the JIT source code was imported from
a submodule and is hence in a new location, requiring updates to the
patches.

Note that 0002-RunTest-adjust-ISO-8859-1-tests-for-Cygwin requires an
unintuitive adjustment in the part where it is re-wrapping the test
data: PCRE2 changed the format of the `Starting code points` section so
that it no longer includes trailing spaces, see
https://github.com/PCRE2Project/pcre2/commit/5bff405f6326#diff-8e816db40d244630db9665260092b29678ae73d55cefc6207d8fce92f3330a00.

Also, the `LICENSE` file now has an `.md` file extension.

Signed-off-by: Matthias Aßhauer <mha1993@live.de>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>