ci(i686): hard-code libgpg-error as a dependency of gnupg
In the i686 variant of the CI builds, we do not have the benefit of
MSYS2's extensive work on the CI definition where it automatically
determines the correct order how packages are supposed to be built.

As a consequence, `gnupg` is being built before `libgpg-error` in
https://github.com/git-for-windows/MSYS2-packages/pull/202 and to make
matters even worse, the latter is not even installed (but needs to be,
so that its `yat2m.exe` can be found -- and used -- by the former).

Since i686 MSYS package builds in Git for Windows land are on their way
out, let's not invest _too_ much energy into a generic solution and
instead simply hard-code the order of `libgpg-error` and `gnupg`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>