msys2-runtime: ignore the newly-added UI tests
In https://github.com/git-for-windows/msys2-runtime/pull/95, I
introduced a new UI-based integration test. We do not need that to build
the package. The test is meant to be run in the `msys2-runtime`
repository's PRs _before_ the patches make it into `MSYS2-packages`.
Therefore it does not make sense to carry it here.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>